Improve the new test adding more asserts to it.
