+++ test/Microsoft.Health.Fhir.Shared.Tests.E2E/Microsoft.Health.Fhir.Shared.Tests.E2E.projitems
@@ -13,7 +13,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)AssemblyValidationsTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rest\AzureStorageBlobHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\ModelExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Rest\BundleAdvancedTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rest\BundleTestsUtil.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rest\BulkDeleteTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rest\BundleEdgeCaseTests.cs" />
